GIT review fixes
