+++ include/Makefile.am
@@ -38,13 +38,15 @@ include_amlutils_HEADERS = \
 			   aml/utils/async.h \
 			   aml/utils/version.h \
 			   aml/utils/macros.h \
-			   aml/utils/features.h \
-			   aml/utils/linux.h
+			   aml/utils/features.h
+
+include_aml_backenddir=$(includedir)/aml/utils/backend
+include_aml_backend_HEADERS = aml/utils/backend/linux.h
 
 if HAVE_HWLOC
 include_aml_area_HEADERS+= aml/area/hwloc.h
 include_aml_replicaset_HEADERS += aml/higher/replicaset/hwloc.h
-include_amlutils_HEADERS += aml/utils/hwloc.h
+include_aml_backend_HEADERS += aml/utils/backend/hwloc.h
 endif
 
 if HAVE_CUDA
@@ -59,5 +61,5 @@ endif
 if HAVE_ZE
 include_aml_area_HEADERS += aml/area/ze.h
 include_aml_dma_HEADERS += aml/dma/ze.h
-include_amlutils_HEADERS += aml/utils/ze.h
+include_aml_backend_HEADERS += aml/utils/backend/ze.h
 endif
