+++ src/Makefile.am
@@ -36,8 +36,9 @@ UTILS_SOURCES = \
 		utils/vector.c \
 		utils/queue.c \
 		utils/async.c \
-		utils/features.c \
-		utils/linux.c
+		utils/features.c
+
+BACKEND_SOURCES = backend/linux.c
 
 HIGHER_SOURCES = \
 		replicaset/replicaset.c \
@@ -49,6 +50,7 @@ LIB_SOURCES = \
 	      $(TILING_SOURCES) \
 	      $(LAYOUT_SOURCES) \
 	      $(UTILS_SOURCES) \
+	      $(BACKEND_SOURCES) \
 	      $(HIGHER_SOURCES) \
 	      aml.c
 
@@ -74,7 +76,7 @@ AM_CPPFLAGS += $(HWLOC_CFLAGS)
 AM_LDFLAGS += $(HWLOC_LIBS)
 libaml_la_SOURCES+=area/hwloc.c
 libaml_la_SOURCES+=replicaset/hwloc.c
-libaml_la_SOURCES+=utils/hwloc.c
+libaml_la_SOURCES+=backend/hwloc.c
 endif
 
 #############################################
@@ -96,5 +98,5 @@ AM_LDFLAGS += $(ZE_LIBS)
 AM_CPPFLAGS += $(ZE_CFLAGS)
 libaml_la_SOURCES+=area/ze.c
 libaml_la_SOURCES+=dma/ze.c
-libaml_la_SOURCES+=utils/ze.c
+libaml_la_SOURCES+=backend/ze.c
 endif
