+++ src/backend/ze.c
@@ -12,35 +12,41 @@
 
 #include "aml/area/ze.h"
 #include "aml/dma/ze.h"
-#include "aml/utils/ze.h"
+#include "aml/utils/backend/ze.h"
 
 int aml_ze_data_create(struct aml_ze_data **out, uint32_t driver_num)
 {
 	uint32_t ze_count = 0;
 
+	// Get driver count.
 	if (zeDriverGet(&ze_count, NULL) != ZE_RESULT_SUCCESS)
 		return -AML_ENOTSUP;
 
+	// Collect drivers
 	ze_driver_handle_t driver[ze_count];
 	if (zeDriverGet(&ze_count, driver) != ZE_RESULT_SUCCESS)
 		return -AML_ENOTSUP;
+
+	// Check driver number is valid.
 	if (driver_num >= ze_count)
 		return -AML_EINVAL;
 
+	// Get number of devices
 	uint32_t num_device = 0;
 	if (zeDeviceGet(driver[driver_num], &num_device, NULL) !=
 	    ZE_RESULT_SUCCESS)
 		return -AML_ENOTSUP;
 
+	// Get devices
 	ze_device_handle_t device[num_device];
 	if (zeDeviceGet(driver[driver_num], &num_device, device) !=
 	    ZE_RESULT_SUCCESS)
 		return -AML_ENOTSUP;
 
+	// Get the number of subdevice for each device.
 	uint32_t num_subdevice[num_device];
 	uint32_t tot_subdevice = 0;
 	memset(num_subdevice, 0, sizeof(num_subdevice));
-
 	for (uint32_t i = 0; i < num_device; i++) {
 		if (zeDeviceGetSubDevices(device[i], num_subdevice + i, NULL) !=
 		    ZE_RESULT_SUCCESS)
@@ -48,6 +54,7 @@ int aml_ze_data_create(struct aml_ze_data **out, uint32_t driver_num)
 		tot_subdevice += num_subdevice[i];
 	}
 
+	// Allocate the returned structure.
 	struct aml_ze_data *data;
 	data = AML_INNER_MALLOC_EXTRA(tot_subdevice + num_device, uint32_t,
 	                              (num_device + tot_subdevice) *
@@ -57,11 +64,15 @@ int aml_ze_data_create(struct aml_ze_data **out, uint32_t driver_num)
 	if (data == NULL)
 		return -AML_ENOMEM;
 
+	// Get alloc space for device handles
 	ze_device_handle_t *device_handles = AML_INNER_MALLOC_GET_EXTRA(
 	        data, tot_subdevice + num_device, uint32_t, struct aml_ze_data);
+
+	// Get alloc space for subdevice numbers and memories number.
 	uint32_t *nums =
 	        AML_INNER_MALLOC_GET_ARRAY(data, uint32_t, struct aml_ze_data);
 
+	// Set returned structure fields.
 	data->driver = driver[driver_num];
 	data->num_device = num_device;
 	data->device = device_handles;
@@ -69,6 +80,7 @@ int aml_ze_data_create(struct aml_ze_data **out, uint32_t driver_num)
 	data->num_subdevice = nums;
 	memcpy(data->num_subdevice, num_subdevice, sizeof(num_subdevice));
 
+	// Set subdevice handles and memories number.
 	device_handles += num_device;
 	nums += num_device;
 	for (uint32_t i = 0; i < num_device; i++) {
@@ -112,17 +124,20 @@ struct aml_area *aml_area_ze_device;
 struct aml_area *aml_area_ze_host;
 struct aml_dma *aml_dma_ze_default;
 
-int aml_ze_init(void)
+int aml_backend_ze_init(void)
 {
 	int err = AML_SUCCESS;
 
+	// backend support check.
 	if (!aml_support_backends(AML_BACKEND_ZE))
 		goto error;
 
+	// Pseudo topology initialization.
 	err = aml_ze_data_create(&aml_ze_default_data, 0);
 	if (err != AML_SUCCESS)
 		goto error;
 
+	// Default device area creation.
 	err = aml_area_ze_device_create(&aml_area_ze_device,
 	                                aml_ze_default_data->device[0], 0,
 	                                ZE_DEVICE_MEM_ALLOC_FLAG_BIAS_CACHED,
@@ -130,11 +145,13 @@ int aml_ze_init(void)
 	if (err != ZE_RESULT_SUCCESS)
 		goto error_with_data;
 
+	// Default host area creation
 	err = aml_area_ze_host_create(&aml_area_ze_host,
 	                              ZE_HOST_MEM_ALLOC_FLAG_BIAS_CACHED, 64);
 	if (err != ZE_RESULT_SUCCESS)
 		goto error_with_ze_area_device;
 
+	// Default dma engine creation
 	err = aml_dma_ze_create(&aml_dma_ze_default,
 	                        aml_ze_default_data->device[0], 256);
 	if (err != ZE_RESULT_SUCCESS)
@@ -158,7 +175,7 @@ error:
 	return err;
 }
 
-int aml_ze_finalize(void)
+int aml_backend_ze_finalize(void)
 {
 	if (aml_area_ze_device != NULL)
 		aml_area_ze_destroy(&aml_area_ze_device);
